ChatBot/requirements.txt
pyTelegramBotAPI
python-dotenv
requests
aiosqlite
aiohttp
